UPDATED: Perseus Dark Mode API
This commit goes with updated Perseus Dark Mode API.

UPDATED: PerseusDarkMode lib to v.1.1.3.

UPDATED: PerseusUISystemKit lib to v.1.1.2.
